+++ src/render/backend/stringtoint.cpp
@@ -86,7 +86,7 @@ int StringToInt::lookupId(const QString &str)
         idx = cache.map.value(str, -1);
         if (idx < 0) {
             idx = cache.reverseMap.size();
-            Q_ASSERT(cache.map.size() == cache.reverseMap.size());
+            Q_ASSERT(size_t(cache.map.size()) == cache.reverseMap.size());
             cache.map.insert(str, idx);
             cache.reverseMap.push_back(str);
         }
